% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigR.R
\name{bigR}
\alias{bigR}
\title{Compute R for convolution distribution.}
\usage{
bigR(r, alpha)
}
\arguments{
\item{r}{a vector of cell type specific r values obtained from littleR for gene i, sorted into decreasing order in terms of mu.}

\item{alpha}{a vector of cell type specific dispersions for gene i, sorted into the same order as r.}
}
\value{
A number used to scale zetas.
}
\description{
Compute R for convolution distribution.
}
\examples{
bigR(c(1, 0.75, 0.5, 0.25), 1)

}
