% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaK.R
\name{deltaK}
\alias{deltaK}
\title{Calculate deltaK.}
\usage{
deltaK(K, zeta)
}
\arguments{
\item{K}{an integer value.}

\item{zeta}{a vector of zeta values computed for each integer from 1 to K.}
}
\value{
A value deltaK such that E(K) = R*deltaK.
}
\description{
Calculate deltaK.
}
\examples{
deltaK(K1, zeta1)

}
