% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littleR.R
\name{littleR}
\alias{littleR}
\title{Compute r values used to calculate R.}
\usage{
littleR(mu, alpha)
}
\arguments{
\item{mu}{a vector of cell type specific means for gene i, sorted into decreasing order.}

\item{alpha}{a vector of cell type specific dispersions for gene i, sorted into the same order as mu.}
}
\value{
A vector of r values for each cell type.
}
\description{
Compute r values used to calculate R.
}
\examples{
mu1 <- c(1,3,2)
alpha1 <- c(2,1,0.5)
littleR(mu1[order(mu1, decreasing = TRUE)], alpha1[order(mu1, decreasing = TRUE)])

}
