% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetaK.R
\name{zetaK}
\alias{zetaK}
\title{Compute zeta values used to calculate delta.}
\usage{
zetaK(r, i, alpha)
}
\arguments{
\item{r}{a vector of cell type specific r values obtained from littleR for gene i, sorted into decreasing order in terms of mu.}

\item{i}{scaling index.}

\item{alpha}{a vector of cell type specific dispersions for gene i, sorted into the same order as mu.}
}
\value{
A number.
}
\description{
Compute zeta values used to calculate delta.
}
\examples{
zetaK(r1, i1, alpha1)

}
